Non-AR and AR 360 video players
There are now two 360 video players, just like there are two for 2D videos, one for Non-AR POI and one for AR POI. The one for AR is its own scene, but the other is in the Tour Map Scene. At the time of this commit, I am not sure how well the non-AR 360 player works. Needs more testing.